% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSpectrum.R
\docType{class}
\name{dataSpectrum-class}
\alias{dataSpectrum-class}
\title{Class for particle spectra}
\description{
\code{dataSpectrum} is the main class
to hold particle spectra information, i.e. the number of
particles differential in charge and energy
}
\section{Slots}{

\describe{
\item{\code{depth.g.cm2}}{Depth at which the spectrum is located in g/cm2}

\item{\code{spectrum}}{\code{\link{spectrum}} (i.e. matrix) structure containing the actual particle numbers}
}}
\author{
Steffen Greilich
}
\seealso{
Other Data.structures: \code{\link{dataSpectrum}};
  \code{\link{spectrum}}
}

