% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSpectrum.R
\name{dataSpectrum}
\alias{dataSpectrum}
\title{Constructor for \code{\link{dataSpectrum}} class}
\usage{
dataSpectrum(SPC.data, depth.g.cm2)
}
\arguments{
\item{SPC.data}{dataSPC object holding spectrum information for multiple sample depths}

\item{depth.g.cm2}{Depth at which the spectrum is located in g/cm2. This can be any
depth, spectra will be interpolated between the adjacent sample depths.}
}
\description{
Constructs a dataSpectrum object from a given dataSPC object,
returning the spectrum at a specific depth
}
\details{
Interpolation is done by linear mixture of the spectra which can lead
to funny shapes such as double peaks. This is, however, the only way without
using explicit transport calculation and also employed for example in TRiP98.
}
\author{
Steffen Greilich
}
\seealso{
Other Data.structures: \code{\link{dataSpectrum-class}};
  \code{\link{spectrum}}
}

