% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSpectrum.R
\name{dose.Gy}
\alias{dose.Gy}
\title{Dose for a spectrum}
\usage{
dose.Gy(x, stopping.power.source, target.material)
}
\arguments{
\item{x}{Object of class \code{\link{dataSpectrum}} or list of objects of this class}

\item{stopping.power.source}{Descriptor for source of stopping power data (\code{\link[libamtrack]{stopping.power.source}})}

\item{target.material}{Descriptor for target material (\code{\link[libamtrack]{material.no}})}
}
\description{
(Vector of) dose in Gy
}
\details{
Uses \code{\link[libamtrack]{libamtrack}} stopping power function
}

