% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.multivariat.weighted.histogram.R
\name{get.multivariate.weighted.histogram}
\alias{get.multivariate.weighted.histogram}
\title{Multivariate histogram from weighted variable}
\usage{
get.multivariate.weighted.histogram(x, i, w, x.min, x.max, n.x.bins,
  log = FALSE)
}
\arguments{
\item{x}{continous variable (vector)}

\item{i}{variable to group i (vector of same length)}

\item{w}{optional weighting variable}

\item{x.min}{minimal value (left side for lowest bin)}

\item{x.max}{maximal value (right side for highest bin)}

\item{log}{if true, binning will be done on a logarithmic scale}

\item{n.x.bin}{number of bins}
}
\value{
data frame with bin mids, bin widths and density
}
\description{
Creates histogram for a continous, grouped variable that can be in
addition be weighted.
}
\details{
Bins are left sided. Values outside [x.min, x.max) are disregarded with a warning
}
\author{
Steffen Greilich
}

