% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.FLUKA.phase.space.R
\name{read.FLUKA.phase.space}
\alias{read.FLUKA.phase.space}
\title{Import of phase space files scored with FLUKA}
\usage{
read.FLUKA.phase.space(path, pattern = "*_phase_space",
  particles = c("PROTON", "DEUTERON", "TRITON", "3-HELIUM", "4-HELIUM",
  "HEAVYION"), remove.backscattered = TRUE, Z.max)
}
\arguments{
\item{path}{path containing the phase space data}

\item{pattern}{file pattern}

\item{particles}{list of FLUKA particle names that will be kept, if NULL, all particles will be kept}

\item{remove.backscattered}{if true, particles travelling upstream are removed}

\item{Z.max}{optional upper limit for particle charge}
}
\value{
data frame
}
\description{
Reads in phase space data scored with the mgdraw routine supplied with
this package, can be from multiple runs and multiple boundaries
}
\author{
Steffen Greilich
}

