% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSpectrum.R
\name{spectrum.at.depth.g.cm2}
\alias{spectrum.at.depth.g.cm2}
\title{R function for interpolation of spc files
with depth, was in libamtrack and moved to HITXML Jul15, sgre}
\usage{
spectrum.at.depth.g.cm2(spectra, depth.g.cm2, interpolate = TRUE)
}
\description{
R function for interpolation of spc files
with depth, was in libamtrack and moved to HITXML Jul15, sgre
}

