% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSpectrum.R
\name{spectrum}
\alias{spectrum}
\title{Create empty spectrum}
\usage{
spectrum(nrow)
}
\arguments{
\item{nrow}{Number of rows in the spectrum matrix}
}
\description{
\code{spectrum} return an
empty spectrum, i.e. a matrix with four
columns particle, kinetic energy (mid bin, in MeV.u),
bin width and number of particles in the mid
(absolute, not density)
}
\author{
Steffen Greilich
}
\seealso{
\code{\link{dataSpectrum}}

Other Data.structures: \code{\link{dataSpectrum-class}};
  \code{\link{dataSpectrum}}
}

