% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSpectrum.R
\name{total.n.particles}
\alias{total.n.particles}
\title{Total number of particles in a spectrum}
\usage{
total.n.particles(x)
}
\arguments{
\item{x}{Object of class \code{\link{dataSpectrum}} or list of objects of this class}
}
\description{
Return sum of particles
}

